reverted changes to NestedObject type, TS was unable to compile
