+++ packages/slice/src/index.ts
@@ -4,4 +4,4 @@ export { produce as createNextState } from 'immer';
 export { default as memoize } from 'memoize-state';
 export * from './slice';
 export * from './types';
-export { makeTypeSafeSelector } from './selector';
+// export { makeTypeSafeSelector } from './selector';
