+++ packages/slice/src/tests/selector.test.ts
@@ -2,7 +2,7 @@
 import { makeGetter, makeTypeSafeSelector, get } from '../selector';
 
 describe('makeGetter/makeTypeSafeSelector', () => {
-  const property = ['property', 'fdf'];
+  const property: ['property', 'fdf'] = ['property', 'fdf'];
   const deepState = {
     some: {
       very: {
@@ -42,20 +42,20 @@ describe('makeGetter/makeTypeSafeSelector', () => {
     'nested',
     'property',
   )<string[]>();
-  const getter1B = makeGetter('another', 'deeply', 'nested', 'property', 0);
+  const getter1B = makeGetter('another', 'deeply', 'nested', 'property', '0');
   const getter1C = makeTypeSafeSelector(
     'another',
     'deeply',
     'nested',
     'property',
-    0,
+    '0',
   )<string>();
   const getter1D = makeTypeSafeSelector(
     'another',
     'deeply',
     'nested',
     'property',
-    0,
+    '0',
   ).bindToInput<typeof deepState>();
   const getter2 = makeGetter(
     'anotherdf',
