+++ packages/slice/src/types.ts
@@ -39,7 +39,7 @@ type IncreaseNum<N extends number> = N extends 0
   ? 11
   : never;
 
-export type NestedObject<
+/* export type NestedObject<
   S extends (string | number)[] | ReadonlyArray<string | number>,
   Start extends number,
   Fin,
@@ -49,16 +49,23 @@ export type NestedObject<
   : Start extends Max
   ? Fin
   : S[Start] extends number
-  ? // ? S[IncreaseNum<Start>] extends string
-    //   ? (NestedObject<S, IncreaseNum<Start>, Fin, Max> &
-    //       Record<string, unknown>)[]
-    //   : NestedObject<S, IncreaseNum<Start>, Fin, Max>
-    S[IncreaseNum<Start>] extends string
+  ? S[IncreaseNum<Start>] extends string
     ? (NestedObject<S, IncreaseNum<Start>, Fin, Max> & NotEmptyObject)[]
     : NestedObject<S, IncreaseNum<Start>, Fin, Max>[]
   : {
       [K in S[Start]]: NestedObject<S, IncreaseNum<Start>, Fin, Max>;
-    };
+    }; */
+
+export type NestedObject<
+  S extends string[] | ReadonlyArray<string>,
+  Start extends number,
+  Fin,
+  Max extends number = GetArrayLength<S>,
+> = number extends Start
+  ? never
+  : Start extends Max
+  ? Fin
+  : { [K in S[Start]]: NestedObject<S, IncreaseNum<Start>, Fin, Max> };
 
 // const g: NestedObject<['hello', 'hi', 0, 'hello', 0, 'supP'], 0, any> = {
 //   hello: { hi: [{ hello: [{ supP: '' }], hi: '' }] },
@@ -102,119 +109,3 @@ export type Getter<
   : GetArrayLength<P> extends 10
   ? O[P[0]][P[1]][P[2]][P[3]][P[4]][P[5]][P[6]][P[7]][P[8]][P[9]]
   : never;
-
-type ANY = string | boolean | number | symbol | object | null | undefined;
-
-export function constArr<P0 extends ANY>(a: [P0]): typeof a;
-export function constArr<P0 extends ANY, P1 extends ANY>(a: [P0, P1]): typeof a;
-export function constArr<P0 extends ANY, P1 extends ANY, P2 extends ANY>(
-  a: [P0, P1, P2],
-): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
->(a: [P0, P1, P2, P3]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
->(a: [P0, P1, P2, P3, P4]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
-  P7 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6, P7]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
-  P7 extends ANY,
-  P8 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6, P7, P8]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
-  P7 extends ANY,
-  P8 extends ANY,
-  P9 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6, P7, P8, P9]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
-  P7 extends ANY,
-  P8 extends ANY,
-  P9 extends ANY,
-  P10 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
-  P7 extends ANY,
-  P8 extends ANY,
-  P9 extends ANY,
-  P10 extends ANY,
-  P11 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11]): typeof a;
-export function constArr<
-  P0 extends ANY,
-  P1 extends ANY,
-  P2 extends ANY,
-  P3 extends ANY,
-  P4 extends ANY,
-  P5 extends ANY,
-  P6 extends ANY,
-  P7 extends ANY,
-  P8 extends ANY,
-  P9 extends ANY,
-  P10 extends ANY,
-  P11 extends ANY,
-  P12 extends ANY,
->(a: [P0, P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12]): typeof a;
-export function constArr<P0 extends ANY[]>(a: P0) {
-  return a;
-}
