+++ packages/slice/tsconfig.build.json
@@ -1,11 +1,12 @@
 {
-    // "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "emitDeclarationOnly": true,
-        "declaration": true,
-        "declarationDir": "./lib",
-        "baseUrl": "./",
-        "lib": ["esnext"]
-    },
-    "include": ["./src/index.ts"]
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "emitDeclarationOnly": true,
+    "declaration": true,
+    "declarationDir": "./lib",
+    "baseUrl": "./",
+    "lib": ["esnext"]
+  },
+  "include": ["./src/index.ts"]
 }
