+++ rollup.config.generator.js
@@ -1,11 +1,14 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import typescript from '@rollup/plugin-typescript';
 import resolve from '@rollup/plugin-node-resolve';
 import babel from '@rollup/plugin-babel';
 import commonjs from '@rollup/plugin-commonjs';
 import replace from '@rollup/plugin-replace';
 import { terser } from 'rollup-plugin-terser';
+
 export default function generateConfig({
   input = './src/index.ts',
+  // eslint-disable-next-line global-require
   pkg = require('./package.json'),
   name = '',
 } = {}) {
@@ -30,6 +33,7 @@ export default function generateConfig({
     plugins: [
       typescript({
         tsconfig: '../../tsconfig.json',
+        declaration: true
       }),
       terser(),
     ],
